Update parse dataset and create database
General update to themodules, and in particular finish with the trip
advisor new dataset.
Some consts string and path handling using the config shared module.
Fix ilp module regarding approval vote with a vote for a candidate the
is not in range of valid candidates. It should not be a problem in
general because we extract only votes that relevant to the chosen
candidates range.