Update interfaces
Update all interfaces in thier usages, such that the configurable
varibles (not consts) from config file, will be in used directly and not
as part of the class.
Furthermore, update the ABC reduction to use candidates ids group and
not range, we get the group from the db given a starting point and size
limit.