Add TGD constraint validation and fix DC special converting cases
Add validation to TGD constraint frontend user input upon button press.

Fix the case where the DC do not use a committee variable (and therefore
all committies are invalid).
Fix the case where the DC does not contain the special relation of
committee and therefore it is not a contextual constraint.